# PoC - optimise memory usage for json serialisation.