@scyzoryck 2022-06-19
After the investigation it looks like the issue is more related to event dispatcher that is leaking memory. Closing this PR.